refactor: Remove unnecessary dependency
