+++ composer.json
@@ -10,13 +10,13 @@
 	],
 	"require": {
 		"php": ">=8.2",
-		"tenantcloud/php-standard": "^2.0",
-		"symfony/var-exporter": "^7.0",
+		"illuminate/support": "^10.0 || ^11.0",
+		"nikic/php-parser": "^5.0",
 		"phpstan/phpdoc-parser": "^2.1",
 		"psr/simple-cache": "^3.0",
 		"symfony/cache": "^7.0",
-		"webmozart/assert": "^1.11",
-		"nikic/php-parser": "^5.0"
+		"symfony/var-exporter": "^7.0",
+		"webmozart/assert": "^1.11"
 	},
 	"require-dev": {
 		"pestphp/pest": "^3.0",
@@ -34,7 +34,10 @@
 	"autoload": {
 		"psr-4": {
 			"GoodPhp\\Reflection\\": "src/"
-		}
+		},
+		"files": [
+			"src/Util/Lazy/functions.php"
+		]
 	},
 	"autoload-dev": {
 		"psr-4": {
@@ -54,6 +57,7 @@
 	"minimum-stability": "stable",
 	"prefer-stable": true,
 	"config": {
+		"sort-packages": true,
 		"allow-plugins": {
 			"pestphp/pest-plugin": true
 		}
